+++ src/utils.ts
@@ -88,11 +88,11 @@ export function globAndPrefix (pattern: string, opts: IOptions | string, prefix:
 }
 
 interface NuxtVersion {
-  name: string;
-  version: string;
-  semver: string;
-  suffix: string;
-  section: string;
+  name: string
+  version: string
+  semver: string
+  suffix: string
+  section: string
 }
 
 export function findNuxtDep (pkg: MutablePackageJson): void | NuxtVersion {
@@ -192,28 +192,36 @@ export function getNuxtConfigName (rootDir: string): string {
   throw new Error(`Can not read nuxt.config from ${rootDir}`)
 }
 
-export async function prepareNodeModules (entrypointPath: string, modulesDir: string): Promise<void> {
+export async function prepareNodeModules (entrypointPath: string, namespaceDir: string): Promise<void> {
   const modulesPath = path.join(entrypointPath, 'node_modules')
 
   try {
-    const prodPath = path.join(entrypointPath, modulesDir)
-    if (fs.existsSync(prodPath)) {
-      consola.log(`Using cached ${modulesDir}`)
-    }
+    const namespacedPath = path.join(entrypointPath, namespaceDir)
     try {
       if (fs.existsSync(modulesPath)) {
         await fs.unlink(modulesPath)
       }
-      await fs.mkdirp(modulesDir)
-    } catch {
-      if (fs.existsSync(prodPath)) {
-        fs.rmdirSync(modulesPath, { recursive: true })
-      } else {
-        fs.moveSync(modulesPath, prodPath)
-      }
+    } catch {}
+    if (fs.existsSync(namespacedPath)) {
+      consola.log(`Using cached ${namespaceDir}`)
+      fs.moveSync(namespaceDir, modulesPath)
     }
-    await fs.symlink(modulesDir, modulesPath)
   } catch (e) {
-    consola.log(`Error linking/unlinking ${modulesDir}.`, e)
+    consola.log(`Error creating ${namespaceDir}.`, e)
+  }
+}
+
+export async function backupNodeModules (entrypointPath: string, namespaceDir: string): Promise<void> {
+  const modulesPath = path.join(entrypointPath, 'node_modules')
+
+  try {
+    const namespacedPath = path.join(entrypointPath, namespaceDir)
+    const stats = await fs.stat(modulesPath)
+    if (stats.isDirectory()) {
+      await fs.rm(namespacedPath, { force: true, recursive: true })
+      await fs.move(modulesPath, namespacedPath)
+    }
+  } catch (e) {
+    consola.log(`Error backing up node_modules to ${namespaceDir}.`, e)
   }
 }